backend/requirements.txt
# Core FastAPI and async support
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Background jobs
celery[redis]==5.3.4
flower==2.0.1
click==8.1.7

# AI and ML
openai==1.3.7
langchain==0.0.350
langgraph==0.0.20
pinecone-client==2.2.4

# External APIs
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
notion-client==2.2.1
PyGithub==2.1.1

# Audio processing (for voice features)
speechrecognition==3.10.0
pydub==0.25.1

# File processing
PyPDF2==3.0.1
python-magic==0.4.27

# Monitoring and logging
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0

# Testing (additional)
pytest-timeout==2.2.0
pytest-mock==3.12.0
hypothesis==6.92.1
faker==20.1.0
aiosqlite==0.19.0
